refactor : persistence 계층에 위치 (#859)
