+++ src/main/java/com/clubber/ClubberServer/domain/calendar/repository/CalendarFilterType.java
@@ -1,4 +1,4 @@
-package com.clubber.ClubberServer.domain.calendar.dto;
+package com.clubber.ClubberServer.domain.calendar.repository;
 
 public enum CalendarFilterType {
     ALL,
